+++ comp2013/src/main/java/comp2013/IView.java
@@ -9,34 +9,7 @@ public interface IView {
      * @param controller Controller to use.
      */
     void setController(SnakeController controller);
-    /**
-     * Gets the current url of the song playing
-     * @return URL of the currently playing song.
-     */
-    String getMusic();
 
-    /**
-     * Set the currently playing music
-     * @param url Location of new file to play ".mp3"
-     */
-    void setMusic(String url);
-
-    /**
-     * Plays the currently loaded song.
-     */
-    void playMusic();
-
-    /**
-     * Overloaded function to play music from a specified timestamp.
-     * @param timeStamp
-     */
-    void playMusic(int timeStamp);
-
-    /**
-     * Stops the music from playing
-     * @return Timestamp of the song that was playing
-     */
-    int stopMusic();
 
     /**
      * Draw the snake again from its new coordinates.