+++ comp2013/src/main/java/comp2013/SnakeMusic.java
@@ -1,47 +1,71 @@
 package comp2013;
 
-import java.io.BufferedInputStream;
-import java.io.FileInputStream;
+import javafx.scene.media.Media;
+import javafx.scene.media.MediaPlayer;
 
-import javazoom.jl.player.Player;
+import java.io.File;
+import javafx.util.Duration;
 
-public class SnakeMusic extends Thread
+public class SnakeMusic implements IMusic
 {
-	private String filename;
-	public Player player;
+	private String M_FileName;
+	private Media M_MusicMedia;
+	private MediaPlayer M_MediaPlayer;
 
 	public SnakeMusic(String filename)
 	{
-		this.filename = filename;
+		this.M_FileName = filename;
+		this.M_MusicMedia = new Media(M_FileName);
+		this.M_MediaPlayer = new MediaPlayer(M_MusicMedia);
 	}
 
-	public void play()
-	{
-		new Thread()
-		{
-			@Override
-			public void run()
-			{
-				super.run();
-				try
-				{
-					//BufferedInputStream buffer = new BufferedInputStream(new FileInputStream(filename));
-					player = new Player(new BufferedInputStream(new FileInputStream(filename)));
-					player.play();
-
-				} catch (Exception e)
-				{
-					System.out.println(e);
-				}
-			}
-		}.start();
+	// Sets the music that you want to play
+	@Override
+	public void setMusic(String filename) {
+		if(filename != null) {
+			// Set the filename.
+			this.M_FileName = filename;
+			// Update the media object.
+			this.M_MusicMedia = new Media(new File(filename).toURI().toString());
+			// Create a new Media Player with the new song.
+			this.M_MediaPlayer = new MediaPlayer(M_MusicMedia);
+			// Play the song.
+			this.M_MediaPlayer.play();
+		}
+		else { this.M_FileName = null; }
 	}
 
 
-
-	public static void getMusicPlay(String filename)
-	{
-		SnakeMusic snakeMusic = new SnakeMusic(filename);
-		snakeMusic.play();
+	// Gets the File Path of the currently playing song.
+	@Override
+	public String getMusic(){
+		return this.M_FileName;
 	}
+
+	// Plays the currently loaded song.
+	@Override
+	public void playMusic() {
+		if(M_MediaPlayer.getStatus() != MediaPlayer.Status.PLAYING){
+			M_MediaPlayer.play();
+		}
+	}
+
+
+	// Overloaded function to play music from a specified timestamp.
+	@Override
+	public void playMusic(int timeStamp) {
+		// Set the number of seconds from the timestamp.
+		Duration seekTime = Duration.seconds(timeStamp);
+		M_MediaPlayer.seek(seekTime);
+		M_MediaPlayer.play();
+	}
+
+	// Stops the music from playing
+	@Override
+	public void stopMusic() {
+		if(M_MediaPlayer.getStatus() == MediaPlayer.Status.PLAYING){
+			M_MediaPlayer.pause();
+		}
+	}
+
 }